Add map_view testing ground
