+++ lib/views/map_view.dart
@@ -1,0 +1,98 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:latlong2/latlong.dart';
+import 'package:homebased_project/models/location.dart';
+
+class MapScreen extends StatefulWidget {
+  const MapScreen({super.key});
+
+  @override
+  State<MapScreen> createState() => _MapScreenState();
+}
+
+class _MapScreenState extends State<MapScreen> {
+  final MapController _mapController = MapController();
+  final Color _markerColor = Color.fromARGB(255, 230, 87, 30);
+
+  // List of locations that are displayed as markers
+
+  final List<Location> _locations = [
+    Location(name: "Clementi", latitude: 1.3162, longitude: 103.7649),
+    Location(name: "National University of Singapore", latitude: 1.2976, longitude: 103.7767),
+  ];
+
+  // Function to add new random marker
+  void _addRandomLocation() {
+    setState(() {
+      _locations.add(Location(
+        name: "Random Place #${_locations.length + 1}",
+        latitude: 1.3162 + (0.01 * _locations.length),
+        longitude: 103.7649 + (0.01 * _locations.length),
+      ));
+    });
+  }
+
+  // Function to add a new location marker
+  void addLocation({Location? location, String? name, double? lat, double? lng}) {
+  setState(() {
+    if (location != null) {
+      _locations.add(location);
+    } else if (name != null && lat != null && lng != null) {
+      _locations.add(Location(name: name, latitude: lat, longitude: lng));
+    } else {
+      // Throw error
+    }
+  });
+}
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("Dynamic Markers Map"),
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.add_location_alt),
+            onPressed: _addRandomLocation,
+          )
+        ],
+      ),
+      body: FlutterMap(
+        mapController: _mapController,
+        options: MapOptions(
+          initialCenter: LatLng(1.3404, 103.7090), // Singapore
+          initialZoom: 13.0,
+        ),
+        children: [
+          TileLayer(
+            urlTemplate: 'https://tile.openstreetmap.org/{z}/{x}/{y}.png',
+            userAgentPackageName: 'com.example.app',
+          ),
+          // Key here is that the markerlayer should be handled by the backend api? or just the location info
+          MarkerLayer(
+            markers: _locations.map((location) {
+              return Marker(
+                point: location.getLatLng(),
+                width: 80,
+                height: 80,
+                child: GestureDetector(
+                  onTap: () {
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      SnackBar(content: Text(location.name)),
+                    );
+                  },
+                  child: Icon(
+                    Icons.location_on,
+                    color: _markerColor,
+                    size: 40,
+                  ),
+                ),
+              );
+            }).toList(),
+          ),
+        ],
+      ),
+    );
+  }
+  }
