+++ lib/views/pages/explore_page.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:homebased_project/views/map_view.dart';
 
 class ExplorePage extends StatelessWidget {
   const ExplorePage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Center(child: Text('Explore Page'));
+    return Center(child: MapScreen());
   }
 }
